Move MarbleEmitterNode testcases to own file (plus fixes in Emitter testcases)
